+++ renderers/lit/src/v0_9/index.ts
@@ -22,4 +22,4 @@ export {A2uiLitElement} from './a2ui-lit-element.js';
 export {A2uiController} from './a2ui-controller.js';
 export {BasicCatalogA2uiLitElement} from './catalogs/basic/index.js';
 export {Context} from './context/context.js';
-export {basicCatalog} from './catalogs/basic/index.js';
+export * from './catalogs/basic/index.js';